+++ CHANGELOG.md
@@ -19,6 +19,8 @@ the main axes is placed on the figure.
 ### Bug Fixes
 
 - Correct logic for VTEC year referenced in links (#534).
+- Fix ambiguous situation around the New Year's that could have two VTEC
+ETNs active at the same time (#533).
 
 ## **1.9.0** (13 Dec 2021)